Refactor SPIDRV_ReInit
